Move common rewriting namers from router-http to namer-core
Several of the io.bouyant.http rewriting namers are inaccessible in namerd
because they are in router-http.

The rewriting namers can simply be moved into namer-core (even though "http" is
in the class package name) because there are no actual HTTP-specific
dependencies.

Fixes #789
